continue sort the syscall decode part and review the add undefined label commit
