+++ sel4_common/src/arch/aarch64/message_info.rs
@@ -1,4 +1,3 @@
-
 #[derive(Eq, PartialEq, Debug, Clone, Copy, PartialOrd, Ord)]
 /// The label of a message.
 pub enum MessageLabel {
@@ -34,7 +33,7 @@ pub enum MessageLabel {
     IRQSetIRQHandler,
     IRQClearIRQHandler,
     DomainSetSet,
-	ARMPageTableMap,
+    ARMPageTableMap,
     ARMPageTableUnmap,
     ARMIOPageTableMap,
     ARMIOPageTableUnmap,
@@ -42,30 +41,30 @@ pub enum MessageLabel {
     ARMPageUnmap,
     ARMPageMapIO,
     ARMPageClean_Data,
-	ARMPageInvalidate_Data,
-	ARMPageCleanInvalidate_Data,
-	ARMPageUnify_Instruction,
-	ARMPageGetAddress,
-	ARMASIDControlMakePool,
-	ARMASIDPoolAssign,
-	ARMVCPUSetTCB,
-	ARMVCPUInjectIRQ,
-	ARMVCPUReadReg,
-	ARMVCPUWriteReg,
-	ARMVCPUAckVPPI,
-	ARMIRQIssueIRQHandlerTrigger,
-	ARMIRQIssueIRQHandlerTriggerCore,
-	ARMSIDIssueSIDManager,
-	ARMSIDGetFault,
-	ARMSIDClearFault,
-	ARMSIDBindCB,
-	ARMSIDUnbindCB,
-	ARMCBIssueCBManager,
-	ARMCBTLBInvalidateAll,
-	ARMCBAssignVspace,
-	ARMCBUnassignVspace,
-	ARMCBTLBInvalidate,
-	ARMCBGetFault,
-	ARMCBClearFault,
+    ARMPageInvalidate_Data,
+    ARMPageCleanInvalidate_Data,
+    ARMPageUnify_Instruction,
+    ARMPageGetAddress,
+    ARMASIDControlMakePool,
+    ARMASIDPoolAssign,
+    ARMVCPUSetTCB,
+    ARMVCPUInjectIRQ,
+    ARMVCPUReadReg,
+    ARMVCPUWriteReg,
+    ARMVCPUAckVPPI,
+    ARMIRQIssueIRQHandlerTrigger,
+    ARMIRQIssueIRQHandlerTriggerCore,
+    ARMSIDIssueSIDManager,
+    ARMSIDGetFault,
+    ARMSIDClearFault,
+    ARMSIDBindCB,
+    ARMSIDUnbindCB,
+    ARMCBIssueCBManager,
+    ARMCBTLBInvalidateAll,
+    ARMCBAssignVspace,
+    ARMCBUnassignVspace,
+    ARMCBTLBInvalidate,
+    ARMCBGetFault,
+    ARMCBClearFault,
     nArchInvocationLabels,
 }
