+++ sel4_common/src/arch/aarch64/mod.rs
@@ -1,7 +1,9 @@
 #![allow(unused)]
 mod arch_tcb;
+mod message_info;
 mod registers;
 pub use arch_tcb::ArchTCB;
+pub use message_info::*;
 pub use registers::*;
 
 pub fn set_timer(timer: usize) {
