+++ src/syscall/invocation/decode/arch/mod.rs
@@ -1,0 +1,2 @@
+pub mod arch_irq_invocation;
+pub mod decode_mmu_invocation;
