+++ src/syscall/invocation/decode/decode_irq_invocation.rs
@@ -8,6 +8,7 @@ use sel4_common::{
 use sel4_cspace::interface::{cte_t, CapTag};
 use sel4_task::{get_currenct_thread, set_thread_state, ThreadState};
 
+use super::arch::arch_irq_invocation::arch_decode_irq_control_invocation;
 use crate::syscall::invocation::invoke_irq::{invoke_clear_irq_handler, invoke_set_irq_handler};
 use crate::{
     config::{irqInvalid, maxIRQ},
@@ -113,55 +114,7 @@ pub fn decode_irq_handler_invocation(label: MessageLabel, irq: usize) -> excepti
     };
 }
 
-fn arch_decode_irq_control_invocation(
-    label: MessageLabel,
-    length: usize,
-    src_slot: &mut cte_t,
-    buffer: Option<&seL4_IPCBuffer>,
-) -> exception_t {
-    if label == MessageLabel::RISCVIRQIssueIRQHandlerTrigger {
-        if length < 4 || get_extra_cap_by_index(0).is_none() {
-            unsafe {
-                current_syscall_error._type = seL4_TruncatedMessage;
-            }
-            return exception_t::EXCEPTION_SYSCALL_ERROR;
-        }
-        let irq = get_syscall_arg(0, buffer);
-        let _trigger = get_syscall_arg(1, buffer) != 0;
-        let index = get_syscall_arg(2, buffer);
-        let depth = get_syscall_arg(3, buffer);
-        let cnode_cap = get_extra_cap_by_index(0).unwrap().cap;
-        let status = check_irq(irq);
-        if status != exception_t::EXCEPTION_NONE {
-            return status;
-        }
-        if is_irq_active(irq) {
-            unsafe {
-                current_syscall_error._type = seL4_RevokeFirst;
-            }
-            debug!("Rejecting request for IRQ {}. Already active.", irq);
-            return exception_t::EXCEPTION_SYSCALL_ERROR;
-        }
-        let lu_ret = lookupSlotForCNodeOp(false, &cnode_cap, index, depth);
-        if lu_ret.status != exception_t::EXCEPTION_NONE {
-            debug!("Target slot for new IRQ Handler cap invalid: IRQ {}.", irq);
-            return lu_ret.status;
-        }
-        set_thread_state(get_currenct_thread(), ThreadState::ThreadStateRestart);
-        invoke_irq_control(
-            irq,
-            convert_to_mut_type_ref::<cte_t>(lu_ret.slot as usize),
-            src_slot,
-        )
-    } else {
-        unsafe {
-            current_syscall_error._type = seL4_IllegalOperation;
-        }
-        return exception_t::EXCEPTION_SYSCALL_ERROR;
-    }
-}
-
-fn check_irq(irq: usize) -> exception_t {
+pub(crate) fn check_irq(irq: usize) -> exception_t {
     if irq > maxIRQ || irq == irqInvalid {
         unsafe {
             current_syscall_error._type = seL4_RangeError;
