+++ src/syscall/invocation/decode/mod.rs
@@ -2,7 +2,7 @@ mod decode_cnode_invocation;
 mod decode_domain_invocation;
 pub mod decode_irq_invocation;
 
-mod decode_mmu_invocation;
+pub mod arch;
 mod decode_tcb_invocation;
 mod decode_untyped_invocation;
 
@@ -23,10 +23,11 @@ use crate::kernel::boot::current_syscall_error;
 use crate::syscall::invocation::decode::decode_irq_invocation::decode_irq_handler_invocation;
 
 use self::{
+    arch::decode_mmu_invocation::decode_mmu_invocation,
     decode_cnode_invocation::decode_cnode_invocation,
     decode_domain_invocation::decode_domain_invocation,
     decode_irq_invocation::decode_irq_control_invocation,
-    decode_mmu_invocation::decode_mmu_invocation, decode_tcb_invocation::decode_tcb_invocation,
+    decode_tcb_invocation::decode_tcb_invocation,
     decode_untyped_invocation::decode_untyed_invocation,
 };
